Change to decimal for more precision.
Stops from rounding up to just 2 decimal places.

+++ database/migrations/2018_03_09_150830_create_grades_table.php
@@ -18,7 +18,7 @@ class CreateGradesTable extends Migration
             $table->integer('assignment_id')->unsigned();
             $table->integer('student_id')->unsigned();
             $table->integer('grader_id')->unsigned();
-            $table->float('grade');
+            $table->decimal('grade', 6, 5);
             $table->string('letter_grade', 5);
             $table->boolean('approved')->default(0);
             $table->timestamps();
